requirements.txt
opencv-python==4.8.1.78
tensorflow==2.13.0
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
matplotlib==3.7.2
seaborn==0.12.2
Pillow==10.0.0
imutils==0.5.4
albumentations==1.3.1
tqdm==4.65.0
RPi.GPIO==0.7.1
adafruit-circuitpython-servokit==1.3.8
flask==2.3.2
flask-cors==4.0.0
